Added utils.py
