+++ scripts/utils.py
@@ -261,7 +261,7 @@ def get_id_to_label(type_):
                        
     return id_to_label
 
-def quat2rot_matrix(q0, q1, q2, q3):
+def quat2rot_matrix(qx, qy, qz, qw):
     """
     Covert a quaternion into a full three-dimensional rotation matrix.
  
@@ -275,19 +275,19 @@ def quat2rot_matrix(q0, q1, q2, q3):
     """
      
     # First row of the rotation matrix
-    r00 = 2 * (q0 * q0 + q1 * q1) - 1
-    r01 = 2 * (q1 * q2 - q0 * q3)
-    r02 = 2 * (q1 * q3 + q0 * q2)
+    r00 = 1 - 2*(qy*qy + qz*qz)
+    r01 = 2 * (qx*qy - qw*qz)
+    r02 = 2 * (qx*qz + qw*qy)
      
     # Second row of the rotation matrix
-    r10 = 2 * (q1 * q2 + q0 * q3)
-    r11 = 2 * (q0 * q0 + q2 * q2) - 1
-    r12 = 2 * (q2 * q3 - q0 * q1)
+    r10 = 2 * (qx*qy + qw*qz)
+    r11 = 1 - 2*(qx*qx + qz*qz)
+    r12 = 2 * (qy*qz - qw*qx)
      
     # Third row of the rotation matrix
-    r20 = 2 * (q1 * q3 - q0 * q2)
-    r21 = 2 * (q2 * q3 + q0 * q1)
-    r22 = 2 * (q0 * q0 + q3 * q3) - 1
+    r20 = 2 * (qy*qz - qw*qy)
+    r21 = 2 * (qy*qz + qw*qx)
+    r22 = 1 - 2*(qx*qx + qy*qy)
      
     # 3x3 rotation matrix
     rot_matrix = numpy.array([[r00, r01, r02],
